// backend/src/domain/profiles/profile.schema.ts
export class ProfileStudent {
  profileImage: string;
  dateOfBirth: string;
  education: string;
  additionalEducation: string;
  averageScore: string;
  fieldOfActivity: string;
  technologyStack: string;
  hardSkills: string;
  softSkills: string;
  сareerObjective: string;
  schedule: string;
  workPlace: string;
  foreignLanguages: string;
  interests: string;
  portfolio: string;
  achievements: string;
}

export class ProfileCompany {
  profileImage: string;
  fieldOfActivity: string;
  learningOpportunity: string;
  aboutCompany: string;
  aboutProjects: string;
  abilityToWorkRemotely: string;
}